casi terminando las transferencias

+++ src/main/java/ar/edu/utn/frbb/tup/model/Cuenta.java
@@ -12,7 +12,7 @@ import ar.edu.utn.frbb.tup.controller.dto.CuentaDto;
 public class Cuenta {
     private long numeroCuenta;
     LocalDateTime fechaCreacion;
-    int balance;
+    double balance;
     TipoCuenta tipoCuenta;
     TipoMoneda moneda;
 
@@ -84,11 +84,11 @@ public class Cuenta {
         return this;
     }
 
-    public int getBalance() {
+    public double getBalance() {
         return balance;
     }
 
-    public Cuenta setBalance(int balance) {
+    public Cuenta setBalance(double balance) {
         this.balance = balance;
         return this;
     }
